fix: only the user can delete or update itself

+++ apps/server/src/modules/users/domain/use-cases/delete-user/index.ts
@@ -1,4 +1,5 @@
 import {
+  ForbiddenException,
   HttpException,
   HttpStatus,
   Injectable,
@@ -9,20 +10,24 @@ import { DeleteUserParamsDto } from '~/modules/users/dto/delete-user.dto';
 import UserRepository from '~/services/database/typeorm/repositories/users-repository';
 import { UseCase } from '~/shared/core/use-case';
 
-type DeleteUserParams = DeleteUserParamsDto;
+type DeleteUserParams = DeleteUserParamsDto & { userId: string };
 type DeleteUserResult = Promise<boolean>;
 
 @Injectable()
 export class DeleteUser implements UseCase<DeleteUserParams, DeleteUserResult> {
   constructor(private readonly userRepository: UserRepository) {}
 
-  public async execute({ id }: DeleteUserParams): DeleteUserResult {
+  public async execute({ id, userId }: DeleteUserParams): DeleteUserResult {
     const user = await this.userRepository.findOneById(id);
     if (!user) {
       throw new HttpException(
         UserUseCaseError.messages.userNotFound(id),
         HttpStatus.NOT_FOUND,
       );
+    }
+
+    if (user?.id?.toValue() !== userId) {
+      throw new ForbiddenException(UserUseCaseError.messages.cannotDeleteUser);
     }
 
     user.delete();
